movement speed is now not frame dependent but time dependent this is better for accuracy across different systems with different capabilities but i still need to make sure that no stuttering is happening as video does not seem as smooth anymore

+++ src/level_builder.cpp
@@ -57,7 +57,7 @@ LevelData LevelBuilder::loadLevel(int levelId)
   }
 
 
-  const int TILE_SIZE = 32;
+  const float TILE_SIZE = 32.0f;
 
   int y = 0;
   for (const auto& row : data["tilemap"])
